requirements.txt
dj-database-url==0.4.0
Django==1.10.3
django-bootstrap3==7.1.0
gunicorn==19.4.5
psycopg2==2.6.2
whitenoise==2.0.6
requests==2.9.1
django-tables2==1.2.6
setuptools==20.10.1
lxml==3.7.2
python-docx==0.8.6
django-crispy-forms==1.6.1
django-searchable-select==1.4.5
selenium==3.0.1
coverage==4.2
python-coveralls==2.9.0
Django-Select2==5.10.0
django-select2-forms==2.0.1
django-autocomplete-light==3.2.7
django_querysetsequence==0.7.2